Seongryeol [성렬 리뷰페이지 중간 커밋]

+++ src/main/java/com/project/foodiefoodie/review/domain/Review.java
@@ -3,12 +3,14 @@ package com.project.foodiefoodie.review.domain;
 import lombok.*;
 
 import java.util.Date;
+import java.util.List;
 
 
 @Getter @Setter @ToString
 @NoArgsConstructor
 @AllArgsConstructor
 public class Review {
+
     private String email;
     private String title;
     private String content;
@@ -16,7 +18,8 @@ public class Review {
     private String businessNo;
     private Date lastUpdated;
     private long likeCnt;
-    private int starRate;
+    private double starRate;
     private String isPrivate;
     private int reportCnt;
+
 }
